requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
asyncpg==0.30.0
boto3==1.34.0
botocore==1.34.162
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crispy-bootstrap5==2025.4
crispy-tailwind==1.0.3
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-crispy-forms==2.4
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.121.2
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-speech==2.33.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httptools==0.7.1
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
jmespath==1.0.1
lxml==5.4.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pefile==2023.2.7
pillow==12.0.0
priority==2.0.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.6.1
pydantic_core==2.41.5
pygame==2.6.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
PyMySQL==1.1.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==7.4.2
reportlab==4.4.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.9.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==12.0
whitenoise==6.7.0
wsproto==1.3.1
exponent_server_sdk
django-storages==1.14.2
boto3==1.34.11